Split utils and constants into seperate files
